+++ nodejs/awsx/ec2/securityGroup.ts
@@ -102,24 +102,28 @@ export class SecurityGroup extends pulumi.ComponentResource {
 utils.Capture(SecurityGroup.prototype).createEgressRule.doNotCapture = true;
 utils.Capture(SecurityGroup.prototype).createIngressRule.doNotCapture = true;
 
-export type SecurityGroupOrId = SecurityGroup | pulumi.Input<string>;
+export type SecurityGroupsOrIds = SecurityGroup[] | pulumi.Input<string[]>;
 
 /** @internal */
 export function getSecurityGroups(
-        vpc: x.ec2.Vpc, name: string, args: SecurityGroupOrId[] | undefined,
+        vpc: x.ec2.Vpc, name: string, args: SecurityGroupsOrIds | undefined,
         opts: pulumi.ResourceOptions | undefined) {
     if (!args) {
         return undefined;
     }
 
-    const result: x.ec2.SecurityGroup[] = [];
+    if (Array.isArray()) {
+        return args;
+    }
+
+    const result: SecurityGroup[] = [];
     for (let i = 0, n = args.length; i < n; i++) {
         const obj = args[i];
-        if (x.ec2.SecurityGroup.isSecurityGroupInstance(obj)) {
+        if (SecurityGroup.isSecurityGroupInstance(obj)) {
             result.push(obj);
         }
         else {
-            result.push(x.ec2.SecurityGroup.fromExistingId(`${name}-${i}`, obj, { vpc }, opts));
+            result.push(SecurityGroup.fromExistingId(`${name}-${i}`, obj, { vpc }, opts));
         }
     }